% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{get_shockinfo}
\alias{get_shockinfo}
\title{Create \code{shockinfo} argument}
\usage{
get_shockinfo(ident="chol", nr_rows=1)
}
\arguments{
\item{ident}{Definition of identification scheme, either \code{chol}, \code{girf} or \code{sign}, or \code{exo}.}

\item{nr_rows}{Number of rows in the created dataframe.}
}
\description{
Creates dummy \code{shockinfo} argument for appropriate use in  \code{irf} function.
}
\details{
Depending on the identification scheme a different \code{shockinfo} argument in the \code{irf} function is needed. To handle this convenient, an appropriate data.frame with is created with this function.
}
\seealso{
\code{\link{irf}}
}
