% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_main.R
\name{IRF.cf}
\alias{IRF.cf}
\title{Counterfactual Analysis}
\usage{
IRF.cf(obj,shockvar,resp,nhor=24,save.store=FALSE)
}
\arguments{
\item{obj}{an object of class \code{bgvar}.}

\item{shockvar}{structural shock of interest.}

\item{resp}{response variable to neutralize.}

\item{nhor}{forecasting horizon.}

\item{save.store}{If set to \code{TRUE} the full posterior is returned. Default is set to \code{FALSE} in order to save storage.}
}
\description{
Function to perform counterfactual analysis. It enables to neutralize the response of a specific variable to a given shock.
}
\examples{
\donttest{
library(BGVAR)
data(eerData)
model.ssvs.eer<-bgvar(Data=eerData,W=W.trade0012,saves=100,burns=100,plag=1,prior="SSVS",
                      eigen=TRUE)
}
# very time-consuming
\dontrun{
IRF.cf(model.ssvs.eer,shockvar="US.stir",resp="US.rer",nhor=24)
}
}
\author{
Maximilian Boeck, Martin Feldkircher
}
