% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{Convert Input List to Matrix}
\usage{
list_to_matrix(datalist)
}
\arguments{
\item{datalist}{A list of length \code{N} which contains each a matrix of size T times k, where T are time periods and k variables per entity.}
}
\value{
Returns a matrix of size T times K (number of time periods times number of total variables).
}
\description{
Converts a list to an appropriate input matrix for use of \code{bgvar}.
}
\details{
Note the naming convention. Columns should indicate entity and variable name, separated by a dot, e.g. \code{US.y}.
}
\seealso{
\code{\link{bgvar}} for estimation of a \code{bgvar} object.
}
\author{
Maximilian Boeck
}
