% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bgvar.R
\name{lps.bgvar.pred}
\alias{lps.bgvar.pred}
\alias{lps}
\title{Compute log-predictive scores}
\usage{
lps(object, ...)
}
\arguments{
\item{object}{an object of class \code{bgvar.predict}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of dimension h times K, whereas h is the forecasting horizon and K is the number of variables in the system.
}
\description{
Computes and prints log-predictive score of an object of class \code{bgvar.predict}.
}
\examples{
\donttest{
set.seed(571)
library(BGVAR)
data(monthlyData)
monthlyData$OC <- NULL
OE.weights <- list(EB=EA.weights) # weights have to have the same name as the country in the data
model.mn <- bgvar(Data=monthlyData,W=W,plag=1,h=8,saves=100,burns=100,prior="MN",
                  OE.weights=OE.weights)
fcast <- predict(model.mn, fhorz=8)
lps   <- lps(fcast)
}
}
\author{
Maximilian Boeck, Martin Feldkircher
}
