% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{matrix_to_list}
\alias{matrix_to_list}
\title{Convert Input Matrix to List}
\usage{
matrix_to_list(datamat)
}
\arguments{
\item{datamat}{A matrix of size T times K, where T are time periods and K total amount of variables.}
}
\value{
returns a list of length \code{N} (number of entities).
}
\description{
Converts a big input matrix to an appropriate list for use of \code{bgvar}.
}
\details{
Note the naming convention. Columns should indicate entity and variable name, separated by a dot, e.g. \code{US.y}.
}
\seealso{
\code{\link{bgvar}} for estimation of a \code{bgvar} object.
}
\author{
Maximilian Boeck
}
