% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvar_main.R
\name{plot.bgvar}
\alias{plot.bgvar}
\title{Plotting function for fitted values}
\usage{
\method{plot}{bgvar}(x, ..., global = TRUE, resp = NULL)
}
\arguments{
\item{x}{an object of class \code{bgvar}.}

\item{...}{additional arguments.}

\item{global}{if \code{TRUE} global fitted values are plotted, otherwise country fitted values.}

\item{resp}{if only a subset of variables or countries should be plotted. If set to default value \code{NULL} all countries/variables are plotted.}
}
\description{
Plots the fitted values in red of either the country VARs or the GVAR (default) along with the original data.
}
\examples{
\donttest{
set.seed(571)
library(BGVAR)
data(monthlyData)
monthlyData$OC <- NULL
OE.weights <- list(EB=EA.weights) # weights have to have the same name as the country in the data
model.mn<-bgvar(Data=monthlyData,W=W,saves=100,burns=100,plag=1,prior="MN",
                  OE.weights=OE.weights)
plot(model.mn, resp="AT")
}
}
