% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.bgvar.CD}
\alias{print.bgvar.CD}
\title{Print convergence diagnostics}
\usage{
\method{print}{bgvar.CD}(x, ...)
}
\arguments{
\item{x}{object of class \code{bgvar.CD}.}

\item{...}{additional arguments.}
}
\value{
No return value.
}
\description{
Print convergence diagnostics
}
