% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvar_main.R
\name{print.bgvar}
\alias{print.bgvar}
\alias{print}
\title{Print bgvar Output}
\usage{
\method{print}{bgvar}(x, ...)
}
\arguments{
\item{x}{an object of class \code{bgvar}.}

\item{...}{other arguments.}
}
\description{
\code{print} prints the main arguments of an \code{bgvar} object.
}
\examples{
\donttest{
library(BGVAR)
data(monthlyData)
monthlyData$OC<-NULL
OE.weights <- list(EB=EA.weights) # weights have to have the same name as the country in the data
model.ssvs<-bgvar(Data=monthlyData,W=W,saves=100,burns=100,plag=1,prior="SSVS",
                  OE.weights=OE.weights)
print(model.ssvs)
}
}
\seealso{
\code{\link{bgvar}} to estimate a \code{bgvar} object.
}
\author{
Maximilian Boeck, Martin Feldkircher
}
