% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bgvar.R
\name{rmse.bgvar.predict}
\alias{rmse.bgvar.predict}
\alias{rmse}
\title{Compute root mean squared errors}
\usage{
rmse(object, ...)
}
\arguments{
\item{object}{an object of class \code{bgvar.predict}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of dimension h times K, whereas h is the forecasting horizon and K is the number of variables in the system.
}
\description{
Computes and prints root mean squared errors (RMSEs) of an object of class \code{bgvar.predict}.
}
\examples{
\donttest{
set.seed(571)
library(BGVAR)
data(monthlyData)
monthlyData$OC <- NULL
OE.weights <- list(EB=EA.weights) # weights have to have the same name as the country in the data
model.mn <- bgvar(Data=monthlyData,W=W,plag=1,h=8,saves=100,burns=100,prior="MN",
                  OE.weights=OE.weights)
fcast <- predict(model.mn, fhorz=8)
rmse  <- rmse(fcast)
}
}
\author{
Maximilian Boeck, Martin Feldkircher
}
