% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvar_main.R
\name{summary.bgvar}
\alias{summary.bgvar}
\alias{summary}
\title{Summarizing Bayesian Global Vector Autoregression Fits}
\usage{
\method{summary}{bgvar}(object, ...)
}
\arguments{
\item{object}{an object of class \code{bgvar}.}

\item{...}{other arguments.}
}
\description{
Output gives model information as well as some descriptive statistics on convergence properties, likelihood, serial autocorrelation in the errors and the average pairwise autocorrelation of cross-country residuals.
}
\examples{
\donttest{
library(BGVAR)
data(monthlyData)
monthlyData$OC<-NULL
OE.weights <- list(EB=EA.weights) # weights have to have the same name as the country in the data
model.ssvs<-bgvar(Data=monthlyData,W=W,saves=100,burns=100,plag=1,prior="SSVS",
                OE.weights=OE.weights,eigen=TRUE)
summary(model.ssvs)
}
}
\seealso{
\code{\link{bgvar}} to estimate a \code{bgvar} object.

\code{\link{avg.pair.cc}} to compute average pairwise cross-country correlation of cross-country residuals separately.

\code{\link{resid.corr.test}} to compute F-test on first-order autocorrelation of cross-country residuals separately.
}
\author{
Maximilian Boeck
}
