% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvar_main.R
\name{vcov.bgvar}
\alias{vcov.bgvar}
\title{Extract variance-covariance matrix}
\usage{
\method{vcov}{bgvar}(object, ..., quantile = 0.5)
}
\arguments{
\item{object}{an object of class \code{bgvar}.}

\item{...}{additional arguments.}

\item{quantile}{reported quantiles. Default is set to median.}
}
\description{
Extracts the global variance-covariance matrix for \code{bgvar} for certain quantiles of the posterior distribution.
}
\examples{
\dontshow{
data(eerData)
cN<-c("EA","US","UK")
eerData<-eerData[cN]
W.trade0012<-apply(W.trade0012[cN,cN],2,function(x)x/rowSums(W.trade0012[cN,cN]))
model.ng <- bgvar(Data=eerData,W=W.trade0012,plag=1,saves=100,burns=100)
vcov(model.ng)
}
\donttest{
library(BGVAR)
data(eerData)
model.ng <- bgvar(Data=eerData,W=W.trade0012,plag=1,saves=100,burns=100)
vcov(model.ng)
}
}
