% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGVAR.R
\name{vcov}
\alias{vcov}
\alias{vcov.bgvar}
\title{Extract Variance-covariance Matrix of Bayesian GVAR}
\usage{
\method{vcov}{bgvar}(object, ..., quantile = 0.5)
}
\arguments{
\item{object}{An object of class \code{bgvar}.}

\item{...}{Additional arguments.}

\item{quantile}{Reported quantiles. Default is set to median.}
}
\value{
Returns an \code{q} times \code{K} times \code{K} array of the global variance-covariance matrix, where \code{q} is the number of specified quantiles (this dimension is dropped if \code{q=1}) and  \code{K} the number of endogenous variables.
}
\description{
Extracts the global variance-covariance matrix for \code{bgvar} for certain quantiles of the posterior distribution.
}
\examples{
\donttest{
library(BGVAR)
data(testdata)
model.ng <- bgvar(Data=testdata,W=W.test,plag=1,draws=100,burnin=100)
vcov(model.ng)
}
}
\seealso{
\code{\link{bgvar}} for estimation of a \code{bgvar} object.
}
