% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvar_main.R
\name{fitted.bgvar}
\alias{fitted.bgvar}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{bgvar}(object, ..., global = TRUE)
}
\arguments{
\item{object}{an object of class \code{bgvar}.}

\item{...}{additional arguments.}

\item{global}{if \code{TRUE} global fitted values are returned otherwise country fitted values.}
}
\description{
Extracts the fitted values for \code{bgvar}.
}
\examples{
\dontshow{
data(eerData)
cN<-c("EA","US","UK")
eerData<-eerData[cN]
W.trade0012<-apply(W.trade0012[cN,cN],2,function(x)x/rowSums(W.trade0012[cN,cN]))
model.ng <- bgvar(Data=eerData,W=W.trade0012,plag=1,saves=100,burns=100)
fitted(model.ng)
}
\donttest{
library(BGVAR)
data(eerData)
model.ng <- bgvar(Data=eerData,W=W.trade0012,plag=1,saves=100,burns=100)
fitted(model.ng)
}
}
