% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bgvar.R
\name{plot.bgvar.pred}
\alias{plot.bgvar.pred}
\title{Plot predictions of bgvar}
\usage{
\method{plot}{bgvar.pred}(x, ..., resp = NULL, Cut = 40)
}
\arguments{
\item{x}{an object of class \code{bgvar.predict}.}

\item{...}{additional arguments.}

\item{resp}{specify a variable to plot predictions.}

\item{Cut}{length of series to be plotted before prediction begins.}
}
\description{
Plots the predictions of an object of class \code{bgvar.predict}.
}
\examples{
\donttest{
set.seed(571)
library(BGVAR)
data(monthlyData)
monthlyData$OC <- NULL
OE.weights <- list(EB=EA.weights) # weights have to have the same name as the country in the data
model.mn <- bgvar(Data=monthlyData,W=W,plag=1,saves=100,burns=100,prior="MN",
                    OE.weights=OE.weights)
fcast <- predict(model.mn)
plot(fcast, resp="US.p", Cut=20)
}
}
\author{
Maximilian Boeck, Martin Feldkircher
}
