% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.bgvar.resid}
\alias{plot.bgvar.pred}
\alias{plot.bgvar.irf}
\alias{plot.bgvar.fevd}
\title{Plotting function for residuals}
\usage{
\method{plot}{bgvar.resid}(x, ..., global = TRUE, resp = NULL)

\method{plot}{bgvar.pred}(x, ..., resp = NULL, Cut = 40)

\method{plot}{bgvar.irf}(x, ..., resp = NULL, shock.nr = 1, cumulative = FALSE)

\method{plot}{bgvar.fevd}(x, ..., ts, k.max = 10)
}
\arguments{
\item{x}{an object of class \code{bgvar.fevd}.}

\item{...}{additional arguments.}

\item{global}{if \code{TRUE} global residuals are plotted, otherwise country residuals.}

\item{resp}{specify a variable to plot predictions.}

\item{Cut}{length of series to be plotted before prediction begins.}

\item{shock.nr}{specify shock to be plotted.}

\item{cumulative}{whether cumulative impulse response functions should be plotted. Default is set to \code{FALSE}.}

\item{ts}{specify the decomposed time series to be plotted.}

\item{k.max}{plots the k series with the highest for the decomposition of \code{ts}.}
}
\value{
No return value.

No return value.

No return value.

No return value.
}
\description{
Either plots country-residuals or the global-residuals.

Plots the predictions of an object of class \code{bgvar.predict}.

Plots the predictions of an object of class \code{bgvar.predict}.

Plots the decomposition of a specific time series into selected structural shocks.
}
\examples{
\donttest{
library(BGVAR)
data(eerData)
model.ssvs <- bgvar(Data=eerData,W=W.trade0012,plag=1,draws=100,burnin=100,
                    prior="SSVS")
summary(model.ssvs)
res <- residuals(model.ssvs)
plot(res, resp="EA")
}
\dontshow{
library(BGVAR)
data(eerData)
cN<-c("EA","US","UK")
eerData<-eerData[cN]
W.trade0012<-apply(W.trade0012[cN,cN],2,function(x)x/rowSums(W.trade0012[cN,cN]))
model.ssvs.eer<-bgvar(Data=eerData,W=W.trade0012,draws=100,burnin=100,plag=1,prior="SSVS",
                      eigen=TRUE)
fcast <- predict(model.ssvs.eer,n.ahead=8,save.store=TRUE)
plot(fcast, resp="US.Dp", Cut=20)
}
\donttest{
library(BGVAR)
data(eerData)
model.ssvs.eer<-bgvar(Data=eerData,W=W.trade0012,draws=100,burnin=100,plag=1,prior="SSVS",
                      eigen=TRUE)
fcast <- predict(model.ssvs.eer,n.ahead=8,save.store=TRUE)
plot(fcast, resp="US.Dp", Cut=20)
}
\donttest{
library(BGVAR)
data(eerData)
model.ssvs.eer<-bgvar(Data=eerData,W=W.trade0012,draws=100,burnin=100,plag=1,prior="SSVS",
                      eigen=TRUE)
# US monetary policy shock
shocks<-list();shocks$var="stir";shocks$cN<-"US";shocks$ident="chol";shocks$scal=-100
irf.chol.us.mp<-irf(model.ssvs.eer,shock=shocks,nhor=24)
# plots an impulse response function
plot(irf.chol.us.mp,resp="US.y")
}
\donttest{
library(BGVAR)
data(eerData)
model.ssvs.eer<-bgvar(Data=eerData,W=W.trade0012,draws=100,burnin=100,plag=1,
                      prior="SSVS",thin=1,eigen=TRUE)
                      
# US monetary policy shock
shocks<-list();shocks$var="stir";shocks$cN<-"US";shocks$ident="chol";shocks$scal=-100
irf.chol.us.mp<-irf(obj=model.ssvs.eer,shock=shocks,nhor=48)

# calculates FEVD for variables US.Dp and EA.y
fevd.us.mp=fevd(obj=irf.chol.us.mp,var.slct=c("US.Dp","EA.y"))

plot(fevd.us.mp, ts="US.Dp", k.max=10)
}
}
\seealso{
\code{\link{irf}}

\code{\link{fevd}} \code{\link{gfevd}}
}
\author{
Maximilian Boeck, Martin Feldkircher

Maximilian Boeck, Martin Feldkircher

Maximilian Boeck
}
